Split sketch visualizer into modules
