+++ rust/kcl-lib/src/tooling/sketch_visualizer/snapshots/sketch_visualizer_dof.png
