+++ rust/kcl-lib/src/tooling/sketch_visualizer/snapshots/sketch_visualizer_ids_control_polygons.png
